+++ src/main/java/com/nick/wood/graphics_library/Vertex.java
@@ -3,24 +3,25 @@ package com.nick.wood.graphics_library;
 
 import com.nick.wood.maths.objects.Vec2f;
 import com.nick.wood.maths.objects.Vec3d;
+import com.nick.wood.maths.objects.Vec3f;
 
 public class Vertex {
 
-	private Vec3d pos;
+	private Vec3f pos;
 	private Vec2f textureCoord;
-	private Vec3d normal;
+	private Vec3f normal;
 
-	public Vertex(Vec3d pos, Vec2f textureCoord, Vec3d normal) {
+	public Vertex(Vec3f pos, Vec2f textureCoord, Vec3f normal) {
 		this.pos = pos;
 		this.textureCoord = textureCoord;
 		this.normal = normal;
 	}
 
-	public Vec3d getPos() {
+	public Vec3f getPos() {
 		return pos;
 	}
 
-	public void setPos(Vec3d pos) {
+	public void setPos(Vec3f pos) {
 		this.pos = pos;
 	}
 
@@ -32,11 +33,11 @@ public class Vertex {
 		this.textureCoord = textureCoord;
 	}
 
-	public Vec3d getNormal() {
+	public Vec3f getNormal() {
 		return normal;
 	}
 
-	public void setNormal(Vec3d normal) {
+	public void setNormal(Vec3f normal) {
 		this.normal = normal;
 	}
 }
