+++ src/main/java/com/nick/wood/graphics_library/objects/mesh_objects/MeshObject.java
@@ -1,9 +1,9 @@
 package com.nick.wood.graphics_library.objects.mesh_objects;
 
-import com.nick.wood.maths.objects.Matrix4d;
+import com.nick.wood.maths.objects.Matrix4f;
 
 public interface MeshObject {
 	Mesh getMesh();
-	Matrix4d getRotationOfModel();
+	Matrix4f getRotationOfModel();
 	String getStringToCompare();
 }
